rename bookmarks storage service
